feat: 게시글 추가/수정/삭제 service 추가

+++ lib/screens/mypage/screen_myComment.dart
@@ -77,7 +77,7 @@ class MyComment extends StatelessWidget {
                   return PostBubble(
                     userDetail.photoUrl,
                     userDetail.nickname,
-                    boardId,
+                    boardId as int,
                     board.userId,
                     board.body,
                     board.likeCnt,
